+++ client/src/components/Filter/components/CategoryCollapse/styles.module.scss
@@ -1,0 +1,110 @@
+@import '../../../../styles/variables.scss';
+
+.wrapper {
+  max-width: 295px;
+  border: 1px solid $--grey-border-color;
+  margin: 50px;
+  color: $--grey-text-color;
+}
+
+.header {
+  padding: 20px 20px;
+  display: flex;
+  justify-content: space-between;
+  color: $--grey-text-color;
+}
+
+.categories {
+  margin-top: 10px;
+  font-size: 16px;
+}
+
+.subcategory {
+  display: flex;
+  justify-content: space-between;
+  margin: 13px 20px;
+}
+
+.slider {
+  display: flex;
+  justify-content: center;
+  margin-top: 30px;
+  margin-bottom: 20px;
+}
+
+.range {
+  display: flex;
+  justify-content: space-between;
+}
+
+.button {
+  background-color: white;
+  border-radius: 8px;
+  border: 1px solid $--grey-border-color;
+  padding: 3px 30px;
+}
+
+.subcategory.active {
+  color: black;
+}
+
+.reset {
+  background-color: white;
+  border: 1px solid black;
+  margin: 20px 85px;
+}
+
+@media (max-width: $medium) {
+  .wrapper {
+    max-width: 200px;
+    padding: 10px;
+    font-size: 10px;
+  }
+
+  .categories {
+    font-size: 14px;
+  }
+  .reset {
+    font-size: 10px;
+    margin: 5px 45px;
+  }
+}
+
+@media (max-width: $small) {
+  .wrapper {
+    display: flex;
+    flex-wrap: wrap;
+    width: 100vw;
+    border: none;
+    margin: 0px;
+    font-size: 10px;
+  }
+
+  .header {
+    padding: 5px;
+    font-size: 12px;
+  }
+
+  .categories {
+    font-size: 12px;
+  }
+  .reset {
+    font-size: 10px;
+    width: 80px;
+    align-self: center;
+    margin: 10px;
+  }
+
+  .filterImage {
+    display: none;
+  }
+
+  .filter_header {
+    display: flex;
+    max-width: 200px;
+  }
+}
+
+.collapse_wrapper {
+  position: relative;
+}
